Rename part of test_andes.py to test_andes_mats.py
